requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
torch>=2.0.0
torchvision>=0.15.0
rasterio>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pystac-client>=0.8.0
planetary-computer>=1.0.0
